Adds dotted lines to the base of the 'target' area for jquery sortable
With long lists there is a problem in that the 'out of set' column is
not very deep, so it can be hard to drag to that column. The target area
is set using css min-height, so there is only so high you can make that.
By adding css dotted border, you can at least see the target area.

Better fix would be welcome.