Refactors the site so that only one config file is required
Puts all the config options you need in the example.settings.php file.
When you create the site, editing this one file should get you started.

Have included the settings file, possibly in places it should not go,
and used a slightly dodgy way to find the include path.

The main script.js becomes script.js.php and is printed to the page
now, but I think that's ok.

Peer review of this would be welcome.